[r133136] Make OQContentAnimatingLayer use the 10.6 + iPhone 3.0 CALayer method, +needsDisplayForKey:. Remove our timer for driving animations, but left most of the rest of OQContentAnimatingLayer's convenience helpers.
Updated the existing subclasses to work with this new approach. The extra bit needed is that since we are going through the real CA animation path, our copy of our layer gets made for the presenattion loop via -initLayer:. We need to subclass this to copy the state that'll be used in the drawing method.

[r133145] Add a note about needing to implement -initWithLayer: in some cases.

[r133202] Added OQContentAnimatingLayer to the iPad project.