[r133345] Moved the OFInternalAPI category implementation to the bottom of the file. Merged the Private category into the main class as a class extension. Renamed a few private methods to have leading underscores like the rest of them.
[r133348] shigeyas (on github) noticed that a bunch of our tests fail if you are running with a locale of ja_JP.

The only way I could find to reproduce this (short of setting my system locale) was:

	defaults write "otest-x86_64" AppleLocale en_US
	defaults write "otest-i386" AppleLocale en_US

Having done this, these changes now "fix" this problem by:

  - Forcing en_US in the unit tests
  - Remove OFRelativeDateParser's currentCalendar ivar, implicit setting thereof and possibly zombies.
  - Update OFRelativeDateParser to take an NSCalendar instead of a timeZone and calendarIdentifier.
  - Update OFRelativeDateParser to apply its calendar and that calendar's locale to any NSDateFormatter it creates.

OFRelativeDateParser defaults to gregorian/local tz/current locale in a couple places if you pass a nil calendar.

As an aside, we now assert that NSDateFormatter instances start out with NSDateFormatterBehavior10_4 instead of setting it (since we are linked on 10.6).