[r133308] Working on <bug://bugs/58059> (Inspector revamp)
OIInspectorRegistry populates the inspection set by walking the responder chain and any delegates and asking any appropriate objects to add any inspected objects to the inspection set.

This change relinquishes the responder-chain walking to OAApplication's applyToResponderChain block handler, but continues to explicitly check the delegates of each responder, because normal responder chain target searches does not include all delegates. Since applyToResponderChain does hit most delegates, some will get examined twice, but the existing code was already having to guard against asking the same object multiple times for its inspected objects (via the seenControllers set), so we won't be adding inspected objects twice for any target or delegate.

The block that is applied to the responder chain performs the same action for both the target object and the target object's delegate, so that code is factored out into a block of its own, and separately applied to the target and the target's delegate (if the target has a delegate).

r:tjw