[r138058] <bug://bugs/63479> (Zoom overlay should be positioned relative to user's fingers, rather than in the center of the screen)
Normally, the zoom HUD shows up a ways out from the center of the two touches that are zooming.  But joelr noticed a special case in iWork where the zoom HUD is directly between the two fingers if they are roughly vertical.  That makes sense because at that angle it's hard to guess where the main part of the user's hand is.  Easy fix, with a made-up constant (20-pixel allowable horizontal offset).